+++ test/test_mpi.py
@@ -1,6 +1,8 @@
 # -*- encoding: utf-8 -*-
 # NOTE: all the following tests should be valid also in a serial execution
 
+from tempfile import TemporaryDirectory
+
 import numpy as np
 import astropy.time as astrotime
 import litebird_sim as lbs
@@ -390,6 +392,73 @@ def test_observation_tod_set_blocks():
     assert_det_info()
 
 
+def test_write_hdf5_mpi(tmp_path):
+    start_time = 0
+    time_span_s = 60
+    sampling_hz = 10
+
+    sim = lbs.Simulation(
+        base_path=tmp_path,
+        start_time=start_time,
+        duration_s=time_span_s,
+    )
+
+    det = lbs.DetectorInfo(
+        name="Dummy detector",
+        sampling_rate_hz=sampling_hz,
+        bandcenter_ghz=100.0,
+        quat=[0.0, 0.0, 0.0, 1.0],
+    )
+
+    num_of_obs = 12
+    sim.create_observations(detectors=[det], num_of_obs_per_detector=num_of_obs)
+
+    file_names = lbs.write_observations(
+        sim, subdir_name="tod", file_name_mask="litebird_tod{global_index:04d}.h5"
+    )
+
+    assert len(file_names) == len(sim.observations)
+
+    if lbs.MPI_ENABLED:
+        # Wait that all the processes have completed writing the files
+        lbs.MPI_COMM_WORLD.barrier()
+
+    tod_path = sim.base_path / "tod"
+    files_found = list(tod_path.glob("litebird_tod*.h5"))
+    assert len(files_found) == num_of_obs, (
+        f"{len(files_found)} files found in {tod_path} instead of "
+        + f"{num_of_obs}: {files_found}"
+    )
+    for idx in range(num_of_obs):
+        cur_tod = tod_path / f"litebird_tod{idx:04d}.h5"
+        assert cur_tod.is_file(), f"File {cur_tod} was expected but not found"
+
+
+def test_simulation_random():
+    sim = lbs.Simulation()
+    assert sim.random is not None
+
+    state = sim.random.bit_generator.state
+
+    comm_world = lbs.MPI_COMM_WORLD
+
+    assert state["bit_generator"] == "PCG64"
+    assert state["has_uint32"] == 0
+    assert state["uinteger"] == 0
+
+    # We only check the state of the first four MPI process. It's important
+    # to ensure that they are all different, but there is little sense in
+    # checking *every* process.
+    if comm_world.rank == 0:
+        assert state["state"]["state"] == 24896973052328222577814399574126207392
+    elif comm_world.rank == 1:
+        assert state["state"]["state"] == 158287254809478086677339590508859947181
+    elif comm_world.rank == 2:
+        assert state["state"]["state"] == 133763967953742274472419503117976972596
+    elif comm_world.rank == 3:
+        assert state["state"]["state"] == 233910118701024945237145923486727240452
+
+
 def main():
     test_observation_time()
     test_construction_from_detectors()
@@ -397,6 +466,32 @@ def main():
     test_observation_tod_two_block_time()
     test_observation_tod_two_block_det()
     test_observation_tod_set_blocks()
+    test_simulation_random()
+
+    # It's critical that all MPI processes use the same output directory
+    if lbs.MPI_ENABLED:
+        if lbs.MPI_COMM_WORLD.rank == 0:
+            tmp_dir = TemporaryDirectory()
+            tmp_path = tmp_dir.name
+            lbs.MPI_COMM_WORLD.bcast(tmp_path, root=0)
+        else:
+            tmp_dir = None
+            tmp_path = lbs.MPI_COMM_WORLD.bcast(None, root=0)
+    else:
+        tmp_dir = TemporaryDirectory()
+        tmp_path = tmp_dir.name
+
+    try:
+        test_write_hdf5_mpi(tmp_path)
+    finally:
+        # Now we can remove the temporary directory, but first make
+        # sure that there are no other MPI processes still waiting to
+        # finish
+        if lbs.MPI_ENABLED:
+            lbs.MPI_COMM_WORLD.barrier()
+
+        if tmp_dir:
+            tmp_dir.cleanup()
 
 
 main()