+++ CHANGELOG.md
@@ -1,12 +1,28 @@
 # HEAD
 
+-   **Breaking change** New API for noise module [#151](https://github.com/litebird/litebird_sim/pull/151):
+
+    -   Function `add_noise` has been renamed to `add_noise_to_observations`, and its parameter `noisetype` has been renamed into `noise_type` for consistency with other parameters (**breaking**)
+
+    -   New functions `add_white_noise` and `add_one_over_f_noise` are exported (they were already implemented but were not visible)
+
+    -   Each `Simulation` object creates random number generators (field `Simulation.random`), in a way that is safe even for MPI applications
+
+-   Upgrade NumPy from 1.20 to 1.21, Numba from 0.54 to 0.55, Rich from 6.2 to 11.0 [#152](https://github.com/litebird/litebird_sim/pull/152)
+
+-   Fix issue [#148](https://github.com/litebird/litebird_sim/issues/148)
+
+-   Use a cache to speed up CI builds [PR#147](https://github.com/litebird/litebird_sim/pull/147)
+
+# Version 0.4.0
+
 - **Breaking change** Drop support for Python 3.6, enable Python 3.9 [#136](https://github.com/litebird/litebird_sim/pull/136)
 
 - **Breaking change** Rename keyword `distribute` to `split_list_over_processes` in `Simulation.create_observations` [#110](https://github.com/litebird/litebird_sim/pull/110)
 
 - **Breaking change** Switch to thermodynamic units in the MBS module [#123](https://github.com/litebird/litebird_sim/pull/123)
 
-- Use a cache to speed up CI builds [PR#147](https://github.com/litebird/litebird_sim/pull/147)
+- Functions to write/load TODs to HDF5 files [#139](https://github.com/litebird/litebird_sim/pull/139) 
 
 - Module for simulating hwp systematics (hwp_sys) [PR#117](https://github.com/litebird/litebird_sim/pull/117). The algebra is described in [Giardiello et al.](https://arxiv.org/abs/2106.08031)